requirements.txt
<<<<<<< HEAD
Flask
flask-mysqldb
mysql-connector-python
pip==24.2
setuptools==72.1.0
wheel==0.43.0
=======
pip==24.2
setuptools==72.1.0
wheel==0.43.0
Flask==2.3.2
>>>>>>> origin/main
bcrypt==4.0.1
joblib
scikit-learn==0.24.2
numpy==1.22.4
pandas==2.2.2
gunicorn==21.0.1
scipy==1.13.1
<<<<<<< HEAD
=======

>>>>>>> origin/main
